Added BEDES Verification flows.  Bug fixes.

+++ app/actors/validators/basic/WithinRange.scala
@@ -49,8 +49,8 @@ case class WithinRange(guid: String,
     Future {
       value.flatMap{ v => v.usage }.flatMap { v =>
         arguments.map { arg =>
-          val min = (arg \ "min").asOpt[Long] orElse (arg \ "min").asOpt[String].map(_.toLong)
-          val max = (arg \ "max").asOpt[Long] orElse (arg \ "max").asOpt[String].map(_.toLong)
+          val min = (arg \ "min").asOpt[Double] orElse (arg \ "min").asOpt[String].map(_.toDouble)
+          val max = (arg \ "max").asOpt[Double] orElse (arg \ "max").asOpt[String].map(_.toDouble)
 
           val g = MapValid(true, Option(v.toString))
           log.debug("Arguments: min {} max {} value", min, max, value)
